+++ CommonAdapter/app/src/main/AndroidManifest.xml
@@ -16,6 +16,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+	    <activity android:name=".ListviewActivity"/>
+	    <activity android:name=".RecyclerviewActivity"/>
     </application>
 
 </manifest>
